storage, imagepicker 완성, Homepage-> NavigationScreen, signup페이지 3개 하나로 묶음
