+++ android/local.properties
@@ -1,5 +1,5 @@
-sdk.dir=/Users/choeseongchan/Library/Android/sdk
-flutter.sdk=/Users/choeseongchan/development/flutter_stable
+sdk.dir=/Users/minhyeokkim/Library/Android/sdk
+flutter.sdk=/Users/minhyeokkim/flutter
 flutter.buildMode=debug
 flutter.versionName=1.0.0
 flutter.versionCode=1
