+++ ios/Flutter/Generated.xcconfig
@@ -1,6 +1,6 @@
 // This is a generated file; do not edit or check into version control.
-FLUTTER_ROOT=/Users/choeseongchan/development/flutter_stable
-FLUTTER_APPLICATION_PATH=/Users/choeseongchan/StudioProjects/Tell-Me
+FLUTTER_ROOT=/Users/minhyeokkim/flutter
+FLUTTER_APPLICATION_PATH=/Users/minhyeokkim/Desktop/dlivetomain
 COCOAPODS_PARALLEL_CODE_SIGN=true
 FLUTTER_TARGET=lib/main.dart
 FLUTTER_BUILD_DIR=build
