+++ linux/flutter/ephemeral/.plugin_symlinks/screen_retriever
@@ -1,1 +1,1 @@
-/Users/choeseongchan/.pub-cache/hosted/pub.dev/screen_retriever-0.1.6/
+/Users/minhyeokkim/.pub-cache/hosted/pub.dev/screen_retriever-0.1.6/
